+++ RELEASE_NOTES.md
@@ -27,6 +27,11 @@ RELEASE NOTES
 Logtalk compiler and runtime
 ----------------------------
 
+* ADDED: The object identifier argument in a `uses/2` directive and the
+module identifier argument in a `use_module/2` directive can now be a
+parameter variable when using the directive in a parametric object or
+a parametric category.
+
 * FIXED: Bug in the compilation of module meta-predicate calls in debug mode
 where the module prefix would be missing.